requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python
torch
torchvision
python-multipart
git+https://github.com/facebookresearch/segment-anything.git